backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
supabase==2.6.0
google-generativeai==0.8.3
httpx==0.27.2
loguru==0.7.2

